seouling/requirements.txt
argcomplete==1.9.2
attrs==19.1.0
awscli==1.16.218
base58==0.2.4
boto3==1.9.208
botocore==1.12.208
certifi==2019.6.16
cfn-flip==1.2.1
chardet==3.0.4
Click==7.0
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.4
django-storages==1.7.1
djangorestframework==3.10.2
docutils==0.14
drf-yasg==1.16.1
durationpy==0.5
flex==6.14.1
future==0.16.0
gunicorn==19.9.0
hjson==3.0.1
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.3
jsonpointer==2.0
lambda-packages==0.19.0
MarkupSafe==1.1.1
packaging==19.1
Pillow==6.1.0
placebo==0.9.0
psycopg2-binary==2.8.3
pyasn1==0.4.6
pyparsing==2.4.2
pypi==2.1
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.2
PyYAML==3.12
requests==2.22.0
rfc3987==1.3.8
rsa==3.4.2
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
strict-rfc3339==0.7
toml==0.10.0
tqdm==4.19.1
troposphere==2.5.0
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.3
validate-email==1.3
Werkzeug==0.12
wsgi-request-logger==0.4.6
